requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# HTTP clients
httpx==0.26.0
aiohttp==3.9.1
requests==2.31.0

# WebSocket support
websockets==12.0
python-socketio==5.11.0

# MQTT client
paho-mqtt==1.6.1

# AI/LLM clients
openai==1.10.0
anthropic==0.8.1
google-generativeai==0.3.2

# YAML configuration
PyYAML==6.0.1

# Logging and monitoring
structlog==24.1.0
python-json-logger==2.0.7

# Utilities
python-dateutil==2.8.2
pytz==2024.1

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
flake8==7.0.0
mypy==1.8.0
